feat: Created Dashboard service, controller, and it's interfaces, and moved GetDashboardSummary from 'transaction repo' to 'dashboard repo'
