+++ src/backend/Auren.Application/Interfaces/Repositories/IDashboardRepository.cs
@@ -1,0 +1,13 @@
+﻿using Auren.Application.DTOs.Responses;
+using Auren.Domain.Enums;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Auren.Application.Interfaces.Repositories
+{
+    public interface IDashboardRepository
+    {
+        Task<DashboardSummaryResponse> GetDashboardSummaryAsync(Guid userId, TimePeriod? timePeriod, CancellationToken cancellationToken);
+    }
+}
