+++ src/backend/Auren.Application/Interfaces/Services/IDashboardService.cs
@@ -1,0 +1,14 @@
+﻿using Auren.Application.Common.Result;
+using Auren.Application.DTOs.Responses;
+using Auren.Domain.Enums;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Auren.Application.Interfaces.Services
+{
+    public interface IDashboardService
+    {
+        Task<Result<DashboardSummaryResponse>> GetDashboardSummary(Guid userId, TimePeriod timePeriod = TimePeriod.ThisMonth, CancellationToken cancellationToken = default);
+    }
+}
