+++ src/backend/Auren.Infrastructure/Repositories/DashboardRepository.cs
@@ -1,0 +1,87 @@
+﻿using Auren.Application.DTOs.Responses;
+using Auren.Application.DTOs.Responses.Transaction;
+using Auren.Application.Extensions;
+using Auren.Application.Interfaces.Repositories;
+using Auren.Domain.Enums;
+using Auren.Infrastructure.Persistence;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Auren.Infrastructure.Repositories
+{
+    public class DashboardRepository(AurenDbContext dbContext, ITransactionRepository transactionRepository) : IDashboardRepository
+    {
+        public async Task<DashboardSummaryResponse> GetDashboardSummaryAsync(Guid userId, TimePeriod? timePeriod, CancellationToken cancellationToken)
+        {
+            var (startDate, endDate) = timePeriod switch
+            {
+                TimePeriod.Last3Months => DateTime.Today.GetLast3MonthRange(),
+                TimePeriod.Last6Months => DateTime.Today.GetLast6MonthRange(),
+                TimePeriod.ThisYear => DateTime.Today.GetThisYearRange(),
+                TimePeriod.LastMonth => DateTime.Today.GetLastMonthRange(),
+                TimePeriod.ThisMonth => DateTime.Today.GetCurrentMonthRange(),
+                TimePeriod.AllTime => (DateTime.MinValue, DateTime.Today),
+                _ => (DateTime.MinValue, DateTime.Today)
+            };
+
+            var monthlyData = await dbContext.Transactions
+                .Where(t => t.UserId == userId && t.TransactionDate >= endDate)
+                .GroupBy(t => new
+                {
+                    IsCurrentMonth = t.TransactionDate >= startDate,
+                    t.TransactionType
+                })
+                .Select(g => new
+                {
+                    g.Key.IsCurrentMonth,
+                    g.Key.TransactionType,
+                    Total = g.Sum(t => t.Amount)
+                })
+                .ToListAsync(cancellationToken);
+
+            var currentIncome = monthlyData
+                .FirstOrDefault(x => x.IsCurrentMonth && x.TransactionType == TransactionType.Income)?.Total ?? 0;
+
+            var lastMonthIncome = monthlyData
+                .FirstOrDefault(x => !x.IsCurrentMonth && x.TransactionType == TransactionType.Income)?.Total ?? 0;
+
+            var currentExpense = monthlyData
+                .FirstOrDefault(x => x.IsCurrentMonth && x.TransactionType == TransactionType.Expense)?.Total ?? 0;
+
+            var lastMonthExpense = monthlyData
+                .FirstOrDefault(x => !x.IsCurrentMonth && x.TransactionType == TransactionType.Expense)?.Total ?? 0;
+
+            var currentBalance = await transactionRepository.GetBalanceAsync(userId, startDate, endDate, cancellationToken);
+            var lastMonthBalance = await transactionRepository.GetBalanceAsync(userId, startDate, endDate, cancellationToken);
+
+            var balanceChange = CalculatePercentageChange(currentBalance.Balance, lastMonthBalance.Balance);
+            var incomeChange = CalculatePercentageChange(currentIncome, lastMonthIncome);
+            var expenseChange = CalculatePercentageChange(currentExpense, lastMonthExpense);
+
+            return new DashboardSummaryResponse(
+                TotalBalance: new TransactionMetricResponse(
+                    Amount: currentBalance.Balance,
+                    PercentageChange: balanceChange
+                ),
+                Income: new TransactionMetricResponse(
+                    Amount: currentIncome,
+                    PercentageChange: incomeChange
+                ),
+                Expense: new TransactionMetricResponse(
+                    Amount: currentExpense,
+                    PercentageChange: expenseChange
+                )
+            );
+        }
+
+        private static decimal CalculatePercentageChange(decimal current, decimal previous)
+        {
+            if (previous == 0) return current > 0 ? 100 : 0;
+            var change = ((current - previous) / previous) * 100;
+
+            return Math.Round(change, 2);
+        }
+    }
+}
